Catch exception on indexing
As seen in #3, some issues during indexing could crash the whole server.
This commit prevent such exception from crashing the server.

Adds a check `_assert_index` to ensure that the index is valid before
using it in handles.
Raising a runtime exception during the indexing will break the index
(pointer to nullptr) and will emit error without crash.
The diagnostics are emitted before the indexing, so that errors in code
are provided even in case of crash, helping the user to fix the RTL.