[CORE] First tentative implementation of dispatch
This commit implement a "call dispatcher" for the RPC protocol.
This will allow threading the function call with regards to the
incoming messages management.
It will enable callback management and server-to-client requests
management, or multi-threaded calls further downstream.

This commit is a WIP, no actual change made, before integration into
Diplomat.